java 11, squirtgun 2.0.0-pre3, decrease plugin.yml api version to 1.13, actual version, moved license file to squirtgun project
