+++ settings.gradle.kts
@@ -5,6 +5,5 @@
 rootProject.name = "pronouns"
 include(":pronouns-api")
 include(":pronouns-squirtgun")
-include(":pronouns-papi")
 include(":pronouns-bukkit")
 include(":pronouns-bungee")
